% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMissingParents.R
\name{addMissingParents}
\alias{addMissingParents}
\title{Add missing parents to an adjacency matrix}
\usage{
addMissingParents(a)
}
\arguments{
\item{a}{An adjMatrix object}
}
\value{
A list of adjMatrix objects where all columns sum to either 0 or 2.
}
\description{
Add missing parents to an adjacency matrix
}
\examples{
a = adjMatrix(c(1,0,0,0), sex=c(1,1))
addMissingParents(a)

b = adjMatrix(rbind(rep(1,4), 0,0,0), sex=c(1,1,1,1))
addMissingParents(b)

}
