% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedbuildr-package.R
\docType{package}
\name{pedbuildr-package}
\alias{pedbuildr}
\alias{pedbuildr-package}
\title{pedbuildr: Pedigree Reconstruction}
\description{
Reconstruct pedigrees from genotype data, by optimising the likelihood over all possible pedigrees subject to given restrictions. Tailor-made plots facilitate evaluation of the output. This package is part of the 'pedsuite' ecosystem for pedigree analysis. In particular, it imports 'pedprobr' for calculating pedigree likelihoods and 'forrel' for estimating pairwise relatedness.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/magnusdv/pedbuildr}
  \item Report bugs at \url{https://github.com/magnusdv/pedbuildr/issues}
}

}
\author{
\strong{Maintainer}: Magnus Dehli Vigeland \email{m.d.vigeland@medisin.uio.no} (\href{https://orcid.org/0000-0002-9134-4962}{ORCID})

}
\keyword{internal}
