% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedbuildr-package.R
\docType{package}
\name{pedbuildr}
\alias{pedbuildr}
\title{pedbuildr: Pedigree reconstruction in R}
\description{
Functions for generating all possible pedigrees connecting a given set of individuals
}
