% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trioData}
\alias{trioData}
\title{Reconstruction example with three individuals}
\format{
A matrix with 3 rows and 100 columns. Each entry contains a genotype
in the form a/b.
}
\usage{
trioData
}
\description{
This dataset contains simulated genotypes for 3 individuals at 100 SNP
markers.
}
\examples{
trioData[, 1:10]

x = list(singleton(1), singleton(2), singleton(3)) |>
  setMarkers(alleleMatrix = trioData, locusAttributes ="snp12")

x

}
\keyword{datasets}
