% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tutankhamun}
\alias{Tutankhamun}
\title{Pedigree of Tutankhamun}
\format{
A data frame with 7 rows and 12 columns:
\itemize{
\item \code{id},\code{fid},\code{mid},\code{sex}: Pedigree columns in standard format
\item \code{D13S317}, ...: Genotype columns for 8 markers
}

See help page for \link{trioData} for further details on format.
}
\source{
Hawass et al. \emph{Ancestry and pathology in King Tutankhamun's family}.
Jama (2010).
}
\usage{
Tutankhamun
}
\description{
A reconstructed pedigree of the Egyptian Pharaoh Tutankhamun, with genotypes
for 8 STR markers, as published by Hawass et al.
}
\keyword{datasets}
