% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMissingParents.R
\name{addMissingParents}
\alias{addMissingParents}
\title{Add missing parents to a pedigree adjacency matrix}
\usage{
addMissingParents(a, maxLinearInbreeding = Inf)
}
\arguments{
\item{a}{An adjMatrix object}

\item{maxLinearInbreeding}{A nonnegative integer, or \code{Inf} (default). If this
is a finite number, it disallows mating between pedigree members X and Y if
X is a linear descendant of Y separated by more than the given number. For
example, setting \code{maxLinearInbreeding = 0} forbids mating between
parent-child, grandparent-grandchild, a.s.o. If \code{maxLinearInbreeding = 1}
then parent-child matings are allowed, but not grandparent-grandchild or
higher.}
}
\value{
A list of adjMatrix objects where all columns sum to either 0 or 2.
}
\description{
Add missing parents to a pedigree adjacency matrix
}
\examples{
a = adjMatrix(c(1,0,0,0), sex=c(1,1))
addMissingParents(a)

b = adjMatrix(rbind(rep(1,4), 0,0,0), sex=c(1,1,1,1))
addMissingParents(b)

}
