% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPeds.R
\name{buildPeds}
\alias{buildPeds}
\title{Build all possible pedigrees}
\usage{
buildPeds(
  ids,
  sex,
  age = NULL,
  knownPO = NULL,
  allKnown = FALSE,
  notPO = NULL,
  connected = TRUE,
  maxLinearInb = Inf,
  sexSymmetry = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{ids}{A vector of ID labels.}

\item{sex}{A vector of the same length as \code{ids}, with entries 1 (male) or 2
(female).}

\item{age}{A numeric vector of the same length as \code{ids}. If \code{age[i]} is less
than \code{age[j]}, individual \code{i} will not be an ancestor of individual \code{j}.
The numbers themselves are irrelevant, only the partial ordering. Note that
no interpretation is made about individuals of equal age.

Alternatively, for finer control, age information may be entered as a
character vector of inequalities, e.g., \code{age = c("1>2", "1>3")}.}

\item{knownPO}{A list of pairs of ID labels: The known parent-offspring
pairs.}

\item{allKnown}{A logical. If TRUE, no other pairs than \code{knownPO} will be
assigned as parent-offspring. If FALSE (default), all pairwise combinations
of ids - except those in \code{notPO} - will be treated as potential
parent-offspring pairs.}

\item{notPO}{A list of pairs of ID labels: Pairs known not to be
parent-offspring.}

\item{connected}{A logical. If TRUE (default), only connected pedigrees are
returned.}

\item{maxLinearInb}{A nonnegative integer, or \code{Inf} (default). If this is a
finite number, it disallows mating between pedigree members X and Y if X is
a linear descendant of Y separated by more than the given number. For
example, \code{maxLinearInb = 1} allows parent-child matings, but not
grandparent-grandchild or more distant. Setting \code{maxLinearInb = 0}
disallows all inbreeding of this type.}

\item{sexSymmetry}{A logical. If TRUE (default), pedigrees which are equal
except for the gender distribution of the \emph{added} parents, are regarded as
equivalent, and only one of each equivalence class is returned. Example:
paternal vs. maternal half sibs.}

\item{verbose}{A logical.}
}
\value{
A list of pedigrees.
}
\description{
Build all possible pedigrees
}
\examples{
p = buildPeds(1:3, sex = c(1,2,1), knownPO = list(c(1,3), c(2,3)))
stopifnot(length(p) == 25)
# plotPeds(p)

# By default, one pedigree was removed above.
# Keep all gender combinations (in this case paternal/maternal half sibs):
p2 = buildPeds(1:3, sex = c(1,2,1), knownPO = list(c(1,3), c(2,3)),
              sexSymmetry = FALSE)
stopifnot(length(p2) == 26)
# plotPeds(p2)

# Remove pedigrees with linear inbreeding
p3 = buildPeds(1:3, sex = c(1,2,1), knownPO = list(c(1,3), c(2,3)),
              maxLinearInb = 0)
stopifnot(length(p3) == 7)
# plotPeds(p3)

}
