% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPeds.R
\name{buildPeds}
\alias{buildPeds}
\title{Build all possible pedigrees}
\usage{
buildPeds(ids, sex, knownPO = NULL, allKnown = F, notPO = NULL,
  verbose = F)
}
\arguments{
\item{ids}{A vector of ID labels}

\item{sex}{A vector of the same length as \code{ids}, with entries 1 (male) or 2
(female)}

\item{knownPO}{A list of pairs of ID labels: The known parent-offspring pairs}

\item{allKnown}{A logical. If TRUE, no other pairs than \code{knownPO} will be
assigned as parent-offspring. If FALSE (default), all pairwise combinations
of ids - except those in \code{notPO} - will be treated as potential
parent-offspring pairs.}

\item{notPO}{A list of pairs of ID labels: Pairs known not to be parent-offspring.}

\item{verbose}{A logical}
}
\value{
A list
}
\description{
Build all possible pedigrees
}
\examples{
p = buildPeds(1:3, sex = c(1,1,1), knownPO = list(c(1,2),c(1,3)))

}
