% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedbuildr-package.R
\docType{package}
\name{pedbuildr}
\alias{pedbuildr}
\title{pedbuildr: Pedigree reconstruction}
\description{
Reconstruct small/medium-sized pedigrees from genotype data. The optimisation
works by generating a list of all possible pedigrees subject to various
restrictions, and comparing the likelihoods of these. Tailor-made plots
facilitate evaluation of the output. This package is part of the 'ped suite'
ecosystem for pedigree analysis in R. In particular, it imports 'pedprobr'
for calculating pedigree likelihoods and 'forrel' for estimating pairwise
relatedness.
}
