% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{reconstruct}
\alias{reconstruct}
\title{Pedigree reconstruction}
\usage{
reconstruct(alleleMatrix, loci, pedlist = NULL, founderInb = 0,
  sortResults = T, verbose = T, ...)
}
\arguments{
\item{alleleMatrix}{A matrix with two columns for each marker}

\item{loci}{A list of marker attributes}

\item{pedlist}{A list of pedigrees. If NULL, \code{\link[=buildPeds]{buildPeds()}} is used to
generate a list}

\item{founderInb}{A number in the interval \code{[0,1]}, used as background
inbreeding level in all founders}

\item{sortResults}{A logical. If TRUE, the output is sorted so that the most
likely pedigree comes first}

\item{verbose}{A logical; verbose output or not}

\item{...}{Additional parameters passed on to \code{\link[=buildPeds]{buildPeds()}}}
}
\value{
A list with two entries:
\itemize{
\item \code{pedlist} : A list of pedigrees. This is the same as the input \code{pedlist}, but possibly sorted (if \code{sortResults = TRUE})
\item \code{logliks} : A numerical vector containing the pedigree log-likelihoods
}
}
\description{
Reconstruct the most likely pedigree from genotype data
}
\examples{
\dontrun{ # Requires the `forrel` package

# Simulate genotype data for a trio family (increase N!)
x = forrel::markerSim(nuclearPed(1), N = 5, alleles = 1:3, seed = 123)

# Extract allele matrix and locus attributes (frequencies a.s.o.)
m = getAlleles(x)
loci = lapply(x$markerdata, attributes)

# Reconstruct the most likely pedigree from the data
res = reconstruct(m, loci, sex = c(1, 2, 1), connected = TRUE, genderSym = T)

# Plot the best pedigrees
plotBestPeds(res)

}

}
