% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trioData}
\alias{trioData}
\title{Reconstruction example with three individuals}
\format{
A data frame with 3 rows and 104 columns. The first 4 columns contain
pedigree info:
\itemize{
\item \code{id}: Individual ID
\item \code{fid}: Paternal ID, where 0 means missing father
\item \code{mid}: Maternal ID, where 0 means missing mother
\item \code{sex}: 1 = male; 2 = female
}

The remaining 100 columns contain the genotypes, in the form a/b, where a
and b are the observed alleles (labelled 1 and 2).
}
\usage{
trioData
}
\description{
This dataset contains simulated genotypes for 3 males at 100 SNP markers.
}
\keyword{datasets}
